+++ noodles-cram/src/async/io.rs
@@ -1,6 +1,6 @@
 //! Async CRAM I/O.
 
-mod reader;
+pub mod reader;
 pub mod writer;
 
 pub use self::{reader::Reader, writer::Writer};